requirements.txt
pygame==2.5.2
windows-curses==2.3.3
numpy==1.24.4
matplotlib==3.7.5
pyaudio==0.2.14
pydub==0.25.1
python-vlc==3.0.20123
Flask==3.0.3